pyproject.toml
[project]
name = "product_sourcing_agent"
version = "0.1.0"
description = "product-sourcing-agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai-tools>=1.9.2",
    "crewai[tools]==1.9.2",
    "fpdf2>=2.8.5",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.0",
    "langchain-tavily>=0.2.17",
    "markdown>=3.10.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.53.1",
    "tavily-python>=0.7.20",
]

[project.scripts]
product_sourcing_agent = "product_sourcing_agent.main:run"
run_crew = "product_sourcing_agent.main:run"
train = "product_sourcing_agent.main:train"
replay = "product_sourcing_agent.main:replay"
test = "product_sourcing_agent.main:test"
run_with_trigger = "product_sourcing_agent.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
